amplify/backend/function/documentdbtest76d3ab8a/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
-e ./src
dnspython==2.2.0; python_version >= '3.6' and python_version < '4.0'
pymongo[srv]==4.0.2
python-dotenv==0.19.2
